graphics.py code review complete.
And I'm not really too happy with it in some ways.  It works a lot
better than it did, but indexed verts don't work and making colors
work with it would be screwy.

REALLY I need a 'vertex' structure that everything generates, then
a 'line' structure, THEN everything goes off of that.  That'd make
life a lot less stupid.  THEN I need a way to turn a vert list into
indexed verts, or at least, it might help (it's an optimization, not
a necessity).

THEN I need a unified shader framework, probably using Pyglet's
Groups.  And for that I need to be using Batches properly.

So, with the current model, it's as good as it gets, but the
current model is pretty lame.

Also, added an arcCorners() method that really needs to have
a proper starting point assigned but works okayish apart from
that.