Code review for actor.py and component.py complete.
Honestly not terrible, apart from some holes.

The input component needs to handle jumps with varying
height.

Maybe we should componentify hitpoint tracking, since
we have about 3 basic behaviors there.

In general, it seems like we might really want a
component event system.  It could be pretty simple,
really.

Also checked out shader.py some which looks fine as far
as it goes, but we're gonna need to be able to screw
around with passing uniforms and such in to shaders a lot...

Or vertex attributes that the shaders read and handle?
I dunno.