Doors work.
We now have a second room, which we can actually enter.

Though generating descriptions and such is still a little
fucked up.  We probably need a layer of indirection in
rooms, and a way to actually store said descriptions
and start generating them.

But Starmaze is becoming actually fairly close to
having a functional core feature set.