Well we have falling blocks now, sorta
Though it exposes weaknesses in the PhysicsObj model
(now kludged over but really still present, having
to do with how Shapes/Bodies/Constraints are added
to the Space).
AND weaknesses in the Block code (the Body is not actually
in the center of the Block and that has to get fixed somehow
if we ever want Blocks to move in any way).