Started writing Chunk class for level building.
Then realized that I could rewrite all the Description
classes and, with some Drake-based assistance, got that
put into a decorator.  Involved some refactoring of physicsobj
and actor arguments.

So right now the game runs, working entirely off of
the old-style descriptions; next step will be to switch
to the new-style.