Zones now exist.
They basically have a small amount of shared state that applies
to all rooms in the Zone, currently just backgrounds and (stubbed)
music.  They also contain the rooms, and will eventually be responsible
for organizing the rooms.  This gives them total control of _how_ the
rooms are organized, allowing for per-Zone special cases

Though currently, Gate destinations are just hard-wired.  Need to
replace that with something.