Added disposal code to VertexArray
I mean, it doesn't WORK properly, but it's THERE.
I suspect it has something to do with order of operations;
if you just straight up close the window, then I think the
OpenGL context is being disposed of before the VertexArray is,
so then it calls GL.DeleteBuffer or something and goes "omgwtfbbq"
because there's no context.

The way to handle this is to have loading and disposal of these
scarce resources handled *entirely* through the Resources system,
which then gets told to nuke everything all at once in the GameWindow
OnUnload() method which purposefully gets called before the OpenGL
context goes kerpoof.

But I could be wrong!  Will have to try it.