Cleaned up resource code a lot, added Texture class
Everything that needs disposal should now be IDisposable.
The ResourceLoader should dispose of it all when it is told to cleanup
in World.OnUnload().

The TestRenderer is currently an exception to this rule since it makes
a VertexArray manually; that should get fixed, as should the VertexArray,
once that side of things is a little more solidified.