Figured out a way to avoid loading every Actor down with a Model for StaticModelRenderer, but it's a little goofy.  Need to look at it not at 3 am.
Essentially we use an IStaticRenderable interface as a mix-in to let us specify that an Actor has a Model for us to render.
Downside, it makes StaticModelRenderer slower because every actor it gets has to be checked to see if it implements IStaticRenderable.  The check is pretty quick, but still.

--HG--
branch : json-experimental