Trying to make Renderer's a bit better by adding a RenderSpec thing to each Actor
In principle, the RenderSpec  is just arguments to a Renderer, and also a specification of which Renderer we are using.  The question is, how do we effectively represent it, are the arguments things that apply to the Renderer at load-time or at render-time, how the hell does this fit with the resource loading system, and how the hell do we coax C#'s type system into doing something vaguely what we want.

This SHOULD be simple but I've gotten all tangled up in it somehow, so I'm committing this so I can go back, work on something else, and come back to this later.