The Great Dependency Untangling.
Was not as great as I'd feared.  Basically we just needed to pull the player and any reference to world generation out of the World; those are now parameters it gets upon creation, which makes sense.  This lets us very nicely put the Actor, Component World and WorldGen base classes into Starmaze.Engine, and everything else is properly in Starmaze.Game.

Also did a certain amount of cleanup on various Components and Actors, updating them to the new event system and fixing minor style things.

Closes issue #63.