Start of making vertex layouts more reasonable.
Added all the new vertex stuff I think we need, next step will be to make graphics and drawing code actually use it.

In the end the _real_ gain is the VertexLayout type, because then a Shader can a) look at a chunk of data and ask what kind of vertex it is and what members it has, and b) say "Oh I know how to draw this" or "Oh I don't know how to draw this".

Eventually.  Once we have a way for a Shader to *know* which elements it needs to be able to draw.  We *might* be able to get that from OpenGL.