Ripped out old VertexAttributeArray code and made everything use the new style code.
In the process, thought up a better way of doing it, of course.  Which is currently there but commented out, because I'm sick of doing the exact same thing in slightly different ways.  And I don't want to risk the type changes involved getting out of hand.  Might look at it again when/if you revamp shaders to slot together more nicely.