Texturing now works!
In a rather crude way, but still.  Better handling of multitexturing and such might be nice, as well as an easier way to scale billboards and a less ad-hoc handling of texture units.

A fair amount of that can be pushed up into a Sprite class though.

Also yanked out location directives in shaders, so we could technically fall back to OpenGL 3.2 now.

There's a moderate bit of test code in this which I will yank out in a later commit.