Resolve issue #53
We now have a FollowCam that follows a particular Actor without letting it scroll out of view.  As planned, the ViewManager handles the screen projection and stuff and just aims where the FollowCam tells it.

It's a little kooky still; smoothing out the interaction between ViewManager and FollowCam would be nice, as now they're kinda stuck together with duct tape.  Still, it works!