Actors now have a RenderingState that contains any necessary data on how to render a particular one.
And holy hell but it actually seems to work.  Needs a lot of cleanup, but I'll get to that.

The basic rundown is, each Actor contains a RenderState or some child of it, which contains 1) a reference to what renderer to use to draw that Actor, and 2) references to whatever data is in the Actor that is necessary to draw it.  I feel like this is a very clingy type, but, all that data's necessary one way or another, or else it's not there, and it does strictly limit/define what a Renderer can get its paws on.

This has gotten more difficult than it should be since it's kinda weirded up the Renderer inheritance tree, but in the end it works.  We still do a dynamic cast of some kind or another but only when adding an actor to a render batch (we have render batches now, too!  And the render manager got turned inside out), which happens relatively rarely, so.  I feel like this should be provably unnecessary, but C# seems to disagree.  There's lots of weird type mongling going on here to try to make things "easy", especially in the matter of preloading renderers.  Should go through and try to clean that up some.