(Drawing Text in progress- still need to break it up into stages)
Gave GUI an actor with a BillBoardRendererState attached to it so I could use
BillBoardRenderer's rendering functionality by giving it new textures. And added
a little scaling to so the text is legible.

Renderer.cs - added a scale property in BillBoardRenderState so in
BillboardRender RenderOne method could apply scale to the transform.