Playing with serialization more.
The "annotate all objects marking which properties to set and such" technique superficially works, but is probably not what we want.  It leaves some fixup to be desired and contaminates all our code with messy annotations that distract from what we're actually doing.

The better technique is probably to provide JsonConverter's for all the appropriate types, which will get a little hairy but keep all the hair in one place.  The main issue with that is that JSON.NET's low-level reading-and-writing API is very stream-based, which is a bloody pain in the arse.  It wants to be this way because stream-based processing is potentially very fast and uses little memory, but it's inconvenient 'cause we really just want to grab JSON items at random and do stuff with them.  But if we gotta write a stupid recursive-descent parser, we gotta do it.

It looks like the way to get a dictionary-like json object is with Newtonsoft.Json.Linq.JObject.  That can be handed a JsonReader as well (which includes a BsonReader if we want to make life hard and fast), Sooooo that combined with a JsonConverter might be the Right Way to go about it.