Testing out new serialization code.
Not sure this is the _best_ method, but it will certainly _work_, it just takes a bit of
annoying but easily-understandable boilerplate.  Which conveniently gives us complete control
over what the fuck actually happens when loading/saving stuff, all in one place.  It also
requires lots of public members on classes and such, which is a bit meh, but we'll roll with
it for now.