Tons of poorly-organized but exciting serialization work going on.
One key insight: We don't generally have to turn live objects into json at all.
Though it might be useful.  But this is important 'cause turning an arbitrary
asset like a VertexArray or an Animation into a filename might be tricky,
especially when it doesn't _have_ a filename, being generated dynamically.
So what we might really want is an _identifier_ for all assets...  That might
work, with an IAsset.

Second key insight: Components and RenderState's and such really aren't supposed
to exist separately from Actor's, so when we are creating them we should probably
take an Actor to pass in to the constructor.  That will probably remove a lot of
what is currently rather narsty and hazardous fix-up-after-the-fact code.