Updated custom json.net-style save-load code a little with a few gleaned tricks
The problem with json.net...

The PROBLEM is this:

The thing lets me give it a JsonConverter that will call custom code to load
and save certain types.  With no JsonConverter you can tell it to output a type
field  in the resulting json.  When I load something with no JsonConverter,
it will look at the type field and magically create a new object of that type.
 When I do use the custom converter, it *won't* look at the type field and say
"Oh I should load that with that custom converter that recognizes that type",
and it *won't* magically create a new object of that type.  It will just
create a JObject.  If I use a custom converter for output but read back the
same thing using the default converter, it *also* won't look at the type field.
So as far as I can tell the best way to manage it is to actually do more or
less the same thing that json.net does, ie look at a list of properties and
stuff to save and load, save a type field, but *actually* just fucking write
all the code to do that ourselves.  Which isn't *hard* but would be *easier*
if we could use the fucking library that is fucking supposed to do all that
shit ANYWAY.

So basically, giving it a list of JsonConverter's and a bunch of random-ass
JSON that could be anything that those JsonConverter's knows how to deal
with **just won't work**.