Changed GUIActors to GUITexts, becuase they weren't really actors, just objects that contained an actor for GUIRendering.
Another change. Can now make GUITexts and edit them all from using the GUI Object.
After using GUI.CreateGUIText, a string is returned that is the key to the hashtable
used for storing the GUITexts. For now the keys are simply a string based on the GUITexts
position. ex if a vector of (-2,10) is given the key will be "-210". This system for the key
may change