We can now save and load SpriteRenderState's.
This took a bit of fixing of the semantics of the save/load code in general.

Remember, JToken is just Something In JSON.  JValue is a value... that is, a leaf node on the JSON tree.  A  JObject is a dictionary of key-value pairs.  A JArray is an array.  JValue's, JObject's and JArray's are all JTokens.