So lots of changes, I integrated the Emitter class into the Particle Component because it made more sense to have the list in the component and actually set the velocity in each particle and not just add a velocity in controller.
Overall can emit particles around a circle shape randomly around it, next may make it so particles generate clockwise in a circle shape.

Instead of straight velocity, there is the particles angle (the direction) and the velocity magnitude ( the force).
Particle Component has a EmitShape which is the shape particles will emit by, so far there is a circle.

And the particle color is implemented from the component.