MAJOR API CHANGE, READ THIS
Used to be, each Component would take a reference to its Actor in its constructor.

As I wrote more and more loading and saving code, this got more and more inconvenient.  But we had to make sure each Component has a reference to its Actor before anything tries to do anything to it!

Thing is, Actor had an AddComponent method already, which did some cleanup type stuff already that was necessary for loading and saving.  AND Actor has special slots for the Body and RenderState components which were handled specially.  So I've reorganized it so Components no longer need an Actor in their constructor, and AddComponent/RemoveComponent do all the necessary shuffling around of references and caching of special components and whatever.  So now the Body and RenderState slots, and indeed the Components list, on Actor are read-only.  Everything should be updated now, and everything appears to work.